backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pgvector==0.2.4
alembic==1.12.1
pydantic==2.5.0
python-dotenv==1.0.0
openai==1.3.5
yt-dlp==2023.11.16
langchain==0.0.345
langchain-openai==0.0.2
langgraph==0.0.68
sentence-transformers==2.2.2
numpy==1.24.3
asyncpg==0.29.0
httpx==0.25.2